# Added erf(x) Float64 Julia implementation