@AhmedsElysium 2025-03-31
**Old:**
Float 64
`@benchmark SpecialFunctions.erf(data) setup=(data=6*rand(Float64)-3) samples=1000000`

> BenchmarkTools.Trial: 217729 samples with 1000 evaluations per sample.
>  Range (min … max):   6.300 ns … 283.700 ns  ┊ GC (min … max): 0.00% … 0.00%
>  Time  (median):     29.500 ns               ┊ GC (median):    0.00%
>  Time  (mean ± σ):   21.993 ns ±  13.209 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

Float32
`@benchmark SpecialFunctions.erf(data) setup=(data=6*rand(Float32)-3) samples=1000000`

> BenchmarkTools.Trial: 312732 samples with 1000 evaluations per sample.
>  Range (min … max):   4.300 ns … 125.100 ns  ┊ GC (min … max): 0.00% … 0.00%
>  Time  (median):     19.900 ns               ┊ GC (median):    0.00%
>  Time  (mean ± σ):   15.035 ns ±   7.951 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

**New:**
Float64
`@benchmark erf(data) setup=(data=6*rand(Float64)-3) samples=1000000`

> BenchmarkTools.Trial: 507504 samples with 1000 evaluations per sample.
>  Range (min … max):  5.400 ns … 4.890 μs  ┊ GC (min … max): 0.00% … 0.00%
>  Time  (median):     8.700 ns             ┊ GC (median):    0.00%
>  Time  (mean ± σ):   8.775 ns ± 9.855 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

Float32
`@benchmark Float32(erf(data)) setup=(data=6*rand(Float64)-3) samples=1000000`

> BenchmarkTools.Trial: 526797 samples with 1000 evaluations per sample.
>  Range (min … max):  5.400 ns … 195.500 ns  ┊ GC (min … max): 0.00% … 0.00%
>  Time  (median):     8.800 ns               ┊ GC (median):    0.00%
>  Time  (mean ± σ):   8.521 ns ±   2.236 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%