@AhmedsElysium 2025-03-31
Float32 implementation available, but not faster than Float64 version due to a exp() call.
Float64 version still faster than old Float32.