@AhmedYKadah 2025-04-05
need to clean up polynomial evaluations. 
code also could use more organization