@mschauer 2025-09-16
Then we should also add a test for these