@oscardssmith 2025-09-19
we should probably should test erfc.