@oscardssmith 2025-11-14
Other than missing tests for Inf, looks good to me. @devmotion any further sugestions?