@AhmedYKadah 2025-11-14
Yes, you're correct. The old implementation was not efficient so I redid it.

# Old
**Float64**
`@benchmark erf(data) setup=(data=6*rand(Float64)-3) samples=1000000`

> BenchmarkTools.Trial: 217081 samples with 1000 evaluations per sample.
 Range (min … max):   6.400 ns … 973.200 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     30.100 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   22.028 ns ±  13.612 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

**Float32**
`@benchmark erf(data) setup=(data=6*rand(Float32)-3) samples=1000000`

> BenchmarkTools.Trial: 308269 samples with 1000 evaluations per sample.
 Range (min … max):   4.300 ns … 160.400 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     19.900 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.207 ns ±   8.609 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%
 
 # New
**Float64**
`@benchmark erf(data) setup=(data=6*rand(Float64)-3) samples=1000000`

> BenchmarkTools.Trial: 604748 samples with 1000 evaluations per sample.
 Range (min … max):  4.700 ns … 164.400 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     7.400 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   7.302 ns ±   3.017 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

**Float32**
`@benchmark erf(data) setup=(data=6*rand(Float32)-3) samples=1000000`

> BenchmarkTools.Trial: 641285 samples with 1000 evaluations per sample.
 Range (min … max):  4.200 ns … 204.800 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.900 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.840 ns ±   2.961 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%